server/vqa/requirements.txt
numpy==1.19.5
Pillow==9.2.0
torch==1.12.0
torchvision
